Create main.cpp
A simple program running on an Arduino Uno that calculates and displays values of the Collatz 3N+1 math enigma.  It increments values through values of N starting at 1 and continues until the values converges back to 1 and then incrementing N and repeats.  The running values of N, the values calculated for 3N+1, the current and maximum values obtained and the value of N at both max iterations and max V.   A delay can be introduced to slow the program to better view progress is possible.